tool_calling/pyproject.toml
[project]
name = "tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.16",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.19",
]
